backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
gunicorn>=21.2.0
python-dotenv>=1.0.0
requests>=2.31.0
pydantic>=2.6.0
pydantic-settings>=2.1.0
python-multipart
email-validator
supabase>=2.0.0
psycopg2-binary
python-jose[cryptography]
fpdf2
openpyxl
pandas
slowapi
passlib[bcrypt]
python-json-logger>=2.0.7
httpx>=0.27.0
boto3>=1.34.0



